Added confirm for delete in js

+++ public/assests/js/main.js
@@ -14,19 +14,24 @@ $(function(){
 
 $(document).ready(function() {
     $(".close").on("click", function() {
-        var button = $(this);
-        var id = button.attr("data-target");
-        var token = button.data('token');
+        var agree = confirm("Are you sure?");
+        if (agree) {
+            var button = $(this);
+            var id = button.attr("data-target");
+            var token = button.data('token');
 
-        $.ajax({
-            url: '/delete/' + id,
-            data: {'_token': token},
-            type: 'DELETE',
-            cache: false,
-            async: true,
-            success: function(result) {
-                button.parent().parent().remove();
-            }
-        });
+            $.ajax({
+                url: '/delete/' + id,
+                data: {'_token': token},
+                type: 'DELETE',
+                cache: false,
+                async: true,
+                success: function(result) {
+                    button.parent().parent().remove();
+                }
+            });
+        }
+
+        return false;
     });
 });
